requirements.txt
flask==3.1.2
gunicorn==21.2.0
qrcode[pil]==8.2
Pillow==10.4.0
